rename Test -> TestDetail
